FPII-2388 : Elevation of privilege vulnerability in Framework Listener CVE-2016-3921 A-29831647
The FrameworkListener::onDataAvailable function in libsysutils does not keep track of whether
or not commands have exceeded its read buffer. This could enable command injection to services
spoken to through libsysutils. Since commands that are not null terminated are not processed,
a carefully constructed string could cause the legitimate command to be ignored while a
user-controllable command is executed instead.
The fix is designed to ignore commands that exceed the maximum buffer length.

Change-Id: Ie03f5911d9462f312d8588e9ee678560ba5bbbc3