rootdir: remove libsigchain.so from LD_PRELOAD.
libsigchain.so is a part of ART, which is not in Ubuntu touch. Including
it in LD_PRELOAD makes most (all?) executable in Android container fails
to start.

This is the reason why the Android container fails to load.

(cherry picked from commit 49dfb0e48304daf784322f2188b4023192d902a2)